rework aux worker
