+++ bin/worker
@@ -9,10 +9,10 @@ else
   puts 'DATABASE_URL not set, no pool change needed'
 end
 
-if ENV['AUX_WORKER']
-  # Run all queues (including :auxiliary)
-  exec "echo $DATABASE_URL && GOOD_JOB_QUEUES='*' bundle exec good_job"
-else
-  # Run all queues (excluding :auxiliary)
-  exec "echo $DATABASE_URL && GOOD_JOB_QUEUES='-auxiliary' bundle exec good_job"
+queue = ENV.fetch('HYRAX_ACTIVE_JOB_QUEUE', 'sidekiq')
+case queue
+when 'sidekiq'
+  exec "echo $DATABASE_URL && bundle exec sidekiq"
+when 'good_job'
+  exec "echo $DATABASE_URL && bundle exec good_job start"
 end
